requirements.txt
streamlit==1.31.1
numpy>=1.24.0
plotly>=5.18.0
pandas==2.2.0
openpyxl==3.1.2
authlib>=1.3.2
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
python-dotenv==1.0.1
google-generativeai>=0.3.0
pygsheets>=2.0.6
